transfer getAssetsSubset from polkaswap
